+++ tests/index.test.ts
@@ -411,21 +411,27 @@ test('workspace option', async (context) => {
 
 test('banner and footer option', async (context) => {
   const content = `console.log("Hello, world!")`
-  const { snapshot } = await testBuild({
+  const { fileMap } = await testBuild({
     context,
     files: {
       'index.ts': content,
     },
     options: {
+      dts: true,
       banner: {
-        js: '// banner',
+        js: '// js banner',
+        dts: '// dts banner',
       },
       footer: {
-        js: '// footer',
+        js: '// js footer',
+        dts: '// dts footer',
       },
     },
   })
-  expect(snapshot).toContain('// banner')
-  expect(snapshot).toContain('// footer')
-  expect(snapshot).toContain(content)
+
+  expect(fileMap['index.js']).toContain('// js banner')
+  expect(fileMap['index.js']).toContain('// js footer')
+
+  // expect(fileMap['index.d.ts']).toContain('// dts banner')
+  // expect(fileMap['index.d.ts']).toContain('// dts footer')
 })
